+++ tests/test_nesting.py
@@ -1,28 +1,33 @@
 
 #
-#   tex2txt.py:
 #   - test of nesting for macros
 #   - test of nesting for environments with replacement
 #
 
-from yalafi import tex2txt
+from yalafi import defs, parameters, parser, utils
 
-options = tex2txt.Options(lang='en', char=True)
+parms = parameters.Parameters('en')
+parms.environment_defs.append(
+        defs.Environ(parms, 'table', repl='[Tabelle]', remove=True))
+p = parser.Parser(parms)
 
 def test_nested_macro():
 
-    latex = '\\textcolor{x}{\\textcolor{y}{z}}'
-    plain, nums = tex2txt.tex2txt(latex, options)
+    latex = '\\usepackage{xcolor}\\textcolor{x}{\\textcolor{y}{z}}'
+    toks = p.parse(latex)
+    plain, pos = utils.get_txt_pos(toks)
     assert plain == 'z'
 
 
 def test_nested_table():
 
     latex = '\\begin{table}A\\begin{table}B\\end{table}C\\end{table}'
-    plain, nums = tex2txt.tex2txt(latex, options)
+    toks = p.parse(latex)
+    plain, pos = utils.get_txt_pos(toks)
     assert plain == '\n\n[Tabelle]\n\n'
 
     latex = '\\begin{table}A\\begin{tablx}B\\end{table}C\\end{table}'
-    plain, nums = tex2txt.tex2txt(latex, options)
+    toks = p.parse(latex)
+    plain, pos = utils.get_txt_pos(toks)
     assert plain == '\n\n[Tabelle]\n\nC\n\n'