+++ tests/test_system_macros.py
@@ -45,3 +45,17 @@ def test_renewcommand():
     plain, nums = tex2txt.tex2txt(latex, options)
     assert plain == 'a\n\nyb'
 
+
+def test_phantom():
+
+    latex = r'A\phantom{XYZ}B\phantom{\label{lab}}C'
+    plain, nums = tex2txt.tex2txt(latex, options)
+    assert plain == 'A BC'
+
+    latex = r'A\hphantom{XYZ}B\hphantom{\label{lab}}C'
+    plain, nums = tex2txt.tex2txt(latex, options)
+    assert plain == 'A BC'
+
+    latex = r'A\vphantom{XYZ}B\vphantom{\label{lab}}C'
+    plain, nums = tex2txt.tex2txt(latex, options)
+    assert plain == 'ABC'