+++ yalafi/packages/__init__.py
@@ -1,0 +1,12 @@
+
+#   this list is used by yalafi/tex2txt.py
+#
+load_table = {
+    '*': [
+        'amsmath',
+        'amsthm',
+        'xcolor',
+    ],
+    '': [
+    ],
+}