+++ HISTORY.md
@@ -1,6 +1,23 @@
 Work in progress
 ----------------
+- yalafi
+    - added submodule yalafi.packages, currently with:
+      amsmath, amsthm, xcolor (incomplete initial versions)
+    - changed otion --pyth to --pack
+    - macros \\documentclass and \\usepackage activate package modules
+      from yalafi.packages
+    - removed redundancies in Parameters.math\_space and
+      Parameters.math\_ignore (issue #29)
+    - shifted declarations from yalafi/parameters.py to package modules
+      (issue #29)
+    - closed issue #30 (optional argument for figure environment,
+      missing \\phantom macros)
+    - \\LTmacros renamed to \\LTinput: also may activate packages
+- yalafi.shell
+    - changed option --python-defs to --packages
+    - added option '--add-packages file'
 - added Vim "compiler" ltyc.vim
+- removed file definitions.py
 - README: updated
 
 Version 1.1.1 (2020/05/14)